# Try to XOR as many words as possible in xorbuf APIs