@julek-wolfssl 2023-03-17
Changes made to `XorWords` and `XorWordsOut` to accept `wolfssl_word**` so that we don't have to do pointer arithmetic after the call to these functions. Checked with godbolt.org to see if the compiler outputs different assembly.