@julek-wolfssl 2023-03-17
[Godbolt comparison](https://godbolt.org/#z:OYLghAFBqd5QCxAYwPYBMCmBRdBLAF1QCcAaPECAMzwBtMA7AQwFtMQByARg9KtQYEAysib0QXACx8BBAKoBnTAAUAHpwAMvAFYTStJg1DIApACYAQuYukl9ZATwDKjdAGFUtAK4sGISWakrgAyeAyYAHI%2BAEaYxCAAHKQADqgKhE4MHt6%2B/oGp6Y4CoeFRLLHxSXaYDplCBEzEBNk%2BfgG2mPZFDPWNBCWRMXGJtg1NLbntCmP9YYPlwwkAlLaoXsTI7BzmAMxhyN5YANQmO27T%2BIIAdAin2CYaAIIPj2G0c0cAbqh46EcAGiQAOokdAKADytHQEC8YQIADZJAB9AgAKiOZCOaAY0yOsMEiJR6KYpBeR3JFMpVOpNNpFPxBB2ZhRRwYSxeJgA7FYnpSGUyWXhTjznryKfxiEcIHgTjsACJHDTCo4y05uVnKoWWaxLDEmACsFiF%2BoVBvu8qOTANRoNcuFHM5dqeL2%2BvyOqhI0S8VEh0Nd6HRXqopCxAlx/vRLCYCgA1iH%2BcyCKGvIJ2c7uWT6XCBUnqUKdiK%2BXCEkSadFZQqYcWiUsg/axeTsbiGSW0UcWBWpU2ky2a1HY/XRY9KXgqFKoAzkgRiCja96Tpz1VXBFOZwQlv2Y7rzPqjukAF6YVBUZcI5HriunBUabcZhtUgD06Me72ADDYgixCBqMYURzWBBXEcyhpOk0T0EcRAAuCABKRwAO6ggoQGog%2BmbUgmLKIcQYKdmgKZJg%2Be54Iex6noS66DrSgLECCOEQlCEBSgyFGonOwboXS1IQN2eJwqxG7RnGCFIeyBacZS5bWFeIn0Uc6IHkeJ4seeYmFjSHbSRa2G4QpJFKeRqlUTS%2BGfjuMmKWRKmzsZ5Jck6Q6UhKUqqhaSoFiqsrqqZBCatJlhpsOZbWsapr6uaCosCFtqDvZHJPG8Hz%2BgCwJIREmDwYZRLopivHWWixKkveXElaV5KYUmbIOup5XZomKqxfeznSp27kWJ5aoah5WpWAFEkUqiECosQ7LapYJzhTJg2okwo29VY4npg5Lo/H8HrEEG6WZRGRxBiGvE7Zu8Z1SyPmBVyNV8YIOY0vml29m21LRLZoY4j21ZtlFi2ORSo7jqeq6zkGC5LpO06zpu25mLulnKfxqmXm5t6XZST5HC%2BeBvh%2BSbIN%2ByC/v%2BXiAcBoF4OBmCQag0FwTpyHyWhxUUjRdFgltTFZQVo3wnt/WlTxYbvQS56omx5jwkdyafkRsMc0sam85LSbmRaMv5fL95xc695SRap6tmxdbfZSmm63lH1sZujVBeSzWudempeYrfljRYgW0s9hqhRNEXttFJqxY6HArLQnD6rwfgcFopCoJwbj%2Be1ChrBsFO7DwpAEJoQcrDGID6ho%2BicJI4eZ9HnC8AoID5xnkdB6QcCwEgaAsMkdBxOQlBNy39DxMAXBmIEWCfHgmwAGp4Bl4LJIwnBpzQtAEHEFcQNEJfRGEjQAJ4z7wa/MMQG/gtE2g1NXadN9j4IMLQW816QWBRkY4i3/gxAn3gnyYBXt%2BYKoNRE1sac4SdBLu8aIxBN4eCwCXaceAWDbxWFQAwwAFBjwnlPCOs9ZAiDEOwKQMhBCKBUOoW%2BuguD6EMMYeO%2BgyYV0gCsVAU5MhfwALTgjMOXTob9nAQFcBMPwXAdhBAYOgAYZQKgSE5CkNIGQBB8IkIIgoMiGCiKGPELgkjqi1AEL0cYnhWjyI6F0OoMwVELDURomYciBGjD6KY8R6iViJ3WJsCQwdQ7F1vjHDg7oEjwmYYiI4wBkDICOH3K4ZgpS4EICQE4ZgdhcCWLwauWg5akBznnAuHAi6kAjlHLx5dK7p0zqkkOHB2E5JLvkopNdUkf2IOkZwkggA%3D%3D%3D). This shows that no extra arithmetic is necessary after the loop.