@julek-wolfssl 2023-03-31
Used `xorbuf` API where possible. Left the math libraries and ports alone.